% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{approx_vcov}
\alias{approx_vcov}
\title{Approximate variance-covariance matrix}
\usage{
approx_vcov(fim)
}
\arguments{
\item{fim}{Fisher information matrix.}
}
\value{
Approximate variance-covariance matrix.
}
\description{
Using the Fisher information matrix, compute an approximate
variance-covariance matrix of the estimated model parameters.
}
