% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{curve_variance.gompertz_fit}
\alias{curve_variance.gompertz_fit}
\title{Gompertz fit}
\usage{
\method{curve_variance}{gompertz_fit}(object, x)
}
\arguments{
\item{object}{object of class \code{gompertz_fit}.}

\item{x}{numeric vector at which to evaluate the variance.}
}
\value{
Numeric vector with the variances of the maximum likelihood curve.
}
\description{
Evaluate the variance of the maximum likelihood curve at different predictor
values.
}
