% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{curve_variance.logistic6_fit}
\alias{curve_variance.logistic6_fit}
\title{6-parameter logistic fit}
\usage{
\method{curve_variance}{logistic6_fit}(object, x)
}
\arguments{
\item{object}{object of class \code{logistic6_fit}.}

\item{x}{numeric vector at which to evaluate the variance.}
}
\value{
Numeric vector with the variances of the maximum likelihood curve.
}
\description{
Evaluate the variance of the maximum likelihood curve at different predictor
values.
}
