% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda.R
\name{drda}
\alias{drda}
\title{Fit a parametric model to dose-response data}
\usage{
drda(
  formula,
  data,
  subset,
  weights,
  na.action,
  mean_function = "logistic4",
  lower_bound = NULL,
  upper_bound = NULL,
  start = NULL,
  max_iter = 10000
)
}
\arguments{
\item{formula}{an object of class \code{link[stats]{formula}} (or one that
can be coerced to that class): a symbolic description of the model to be
fitted. Currently supports only formulas of the type \code{response ~ log_dose}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the model. If not found in \code{data}, the variables are taken
from \code{environment(formula)}, typically the environment from which \code{drda}
is called.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting
process. If provided, weighted least squares is used with weights \code{weights}
(that is, minimizing \code{sum(weights * residuals^2)}), otherwise ordinary
least squares is used.}

\item{na.action}{a function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{\link[base]{options}}, and is \code{\link[stats]{na.fail}} if that is
unset. The ‘factory-fresh’ default is \code{\link[stats]{na.omit}}. Another
possible value is \code{NULL}, no action. Value \code{\link[stats]{na.exclude}}
can be useful.}

\item{mean_function}{the model to be fitted. Currently only "logistic4" is
supported.}

\item{lower_bound}{numeric vector of length 4 with the minimum admissible
values of the parameters.}

\item{upper_bound}{numeric vector of length 4 with the maximum admissible
values of the parameters.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}
}
\value{
An object of class "drda", that is a list containing the following
components:
\describe{
\item{converged}{boolean value assessing if the optimization algorithm
converged or not.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean value set to \code{TRUE} if optimization was
constrained.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{sigma}{corrected maximum likelihood estimate of the standard
deviation.}
\item{loglik}{maximum value of the log-likelihood function.}
\item{df.residuals}{residual degrees of freedom.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{(only for weighted fits) the specified weights.}
\item{fisher.info}{Fisher information matrix evaluated at the maximum
likelihood estimator.}
\item{call}{the matched call.}
\item{terms}{the \code{\link[stats]{terms}} object used.}
\item{model}{the model frame used.}
\item{na.action}{(where relevant) information returned by
\code{\link[stats]{model.frame}} on the special handling of \code{NA}s.}
}
}
\description{
Use a Newton trust-region method to fit a curve to dose-response data.
}
