% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda.R
\name{drda_fit_logistic}
\alias{drda_fit_logistic}
\alias{drda_fit_logistic6}
\alias{drda_fit_logistic5}
\alias{drda_fit_logistic4}
\alias{drda_fit_logistic2}
\title{Fit dose-response data}
\usage{
drda_fit_logistic6(x, y, start, max_iter, lower_bound, upper_bound, w)

drda_fit_logistic5(x, y, start, max_iter, lower_bound, upper_bound, w)

drda_fit_logistic4(x, y, start, max_iter, lower_bound, upper_bound, w)

drda_fit_logistic2(x, y, start, max_iter, lower_bound, upper_bound, w)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{lower_bound}{numeric vector with the minimum admissible values of the
parameters.}

\item{upper_bound}{numeric vector with the maximum admissible values of the
parameters.}

\item{w}{an optional vector of weights to be used in the fitting
process.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value (found) of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Use a Newton trust-region method to fit a logistic function to dose-response
data.
}
