% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda-package.R
\docType{package}
\name{drda_package}
\alias{drda_package}
\title{Growth curve fitting and dose-response data analysis}
\description{
\code{drda} is a package for fitting growth curves, with a particular emphasis on
dose-response data analysis.
}
\section{Available functions}{


\itemize{
\item{\code{\link{drda}}: main function for fitting observed data.}
\item{\code{\link{logistic2_fn}}: 2-parameter logistic function.}
\item{\code{\link{logistic4_fn}}: 4-parameter logistic function.}
\item{\code{\link{logistic5_fn}}: 5-parameter logistic function.}
\item{\code{\link{logistic6_fn}}: 6-parameter logistic function.}
\item{\code{\link{gompertz_fn}}: Gompertz function.}
\item{\code{\link{nauc}}: normalized area under the curve.}
\item{\code{\link{naac}}: normalized area above the curve.}
}
}

\references{
Malyutina, A., Tang, J., Pessia, A. drda: An R package for
growth curve fitting and dose-response data analysis. 2020.
}
