% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{effective_dose}
\alias{effective_dose}
\title{Effective dose}
\usage{
effective_dose(object, y, type)
}
\arguments{
\item{object}{fit object as returned by \code{\link[drda]{drda}}.}

\item{y}{numeric vector with response levels.}

\item{type}{character string with either "relative" (default) or "absolute".}
}
\value{
Numeric vector with the effective doses corresponding to each value
of \code{y}.
}
\description{
Estimate effective doses, that is the \code{x} values for which \code{f(x) = y}.
}
\details{
Given a fitted model \verb{f(x; theta)} we seek the values \code{x} at which the
function is equal to the specified response values.

If responses are given on a relative scale (\code{type = "relative"}), then \code{y} is
expected to range in the interval \verb{(0, 1)}.

If responses are given on an absolute scale (\code{type = "absolute"}), then \code{y}
is free to vary on the whole real line. Note, however, that the solution
does not exist when \code{y} is not in the image of the function.
}
