% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{find_optimum}
\alias{find_optimum}
\alias{find_optimum_constrained}
\title{Fit a function to observed data}
\usage{
find_optimum(object)

find_optimum_constrained(object, constraint, known_param)
}
\arguments{
\item{object}{object of some model class.}

\item{constraint}{boolean matrix with constraint specifications.}

\item{known_param}{numeric vector of fixed known parameters.}
}
\value{
A list with the following components:
\describe{
\item{optimum}{maximum likelihood estimates of the model parameters.}
\item{minimum}{(local) minimum of the residual sum of squares around the
means.}
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
}
}
\description{
Use a Newton trust-region method to fit a function to observed data.
}
