% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{fisher_info.gompertz}
\alias{fisher_info.gompertz}
\title{Gompertz fit}
\usage{
\method{fisher_info}{gompertz}(object, theta, sigma)
}
\arguments{
\item{object}{object of class \code{gompertz}.}

\item{theta}{numeric vector with the model parameters.}

\item{sigma}{estimate of the standard deviation.}
}
\value{
Fisher information matrix evaluated at \code{theta}.
}
\description{
Evaluate the Fisher information matrix at the maximum likelihood estimate.
}
\details{
Let \verb{mu(x; theta)} be the Gompertz function. We assume that our observations
\code{y} are independent and such that \verb{y = mu(x; theta) + sigma * epsilon},
where \code{epsilon} has a standard Normal distribution \code{N(0, 1)}.

The 4-by-4 (symmetric) Fisher information matrix is the expected value of
the negative Hessian matrix of the log-likelihood function.
}
