% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{fisher_info.logistic4}
\alias{fisher_info.logistic4}
\title{4-parameter logistic fit}
\usage{
\method{fisher_info}{logistic4}(object, theta, sigma)
}
\arguments{
\item{object}{object of class \code{logistic4}.}

\item{theta}{numeric vector with the model parameters.}

\item{sigma}{estimate of the standard deviation.}
}
\value{
Fisher information matrix evaluated at \code{theta}.
}
\description{
Evaluate the Fisher information matrix at the maximum likelihood estimate.
}
\details{
Let \verb{mu(x; theta)} be the 4-parameter logistic function. We assume that our
observations \code{y} are independent and such that
\verb{y = mu(x; theta) + sigma * epsilon}, where \code{epsilon} has a standard Normal
distribution \code{N(0, 1)}.

The 4-by-4 (symmetric) Fisher information matrix is the expected value of
the negative Hessian matrix of the log-likelihood function.
}
