% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{fit.gompertz}
\alias{fit.gompertz}
\alias{fit_constrained.gompertz}
\title{Gompertz fit}
\usage{
\method{fit}{gompertz}(object)

\method{fit_constrained}{gompertz}(object)
}
\arguments{
\item{object}{object of class \code{gompertz}.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value found of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Fit a Gompertz function to observed data with a Maximum Likelihood approach.
}
\details{
The Gompertz function \verb{f(x; theta)} is defined here as

\verb{alpha + (beta - alpha) exp(-exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate, and \code{phi} related to the function
value at \code{x = 0}.
}
