% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{fit.logistic2}
\alias{fit.logistic2}
\alias{fit_constrained.logistic2}
\title{2-parameter logistic fit}
\usage{
\method{fit}{logistic2}(object)

\method{fit_constrained}{logistic2}(object)
}
\arguments{
\item{object}{object of class \code{logistic2}.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value found of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Fit a 2-parameter logistic function to observed data with a Maximum
Likelihood approach.
}
\details{
The 2-parameter logistic function \verb{f(x; theta)} is defined here as

\code{1 / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(eta, phi)}, \code{eta} is the steepness of the curve or growth
rate (also known as the Hill coefficient), and \code{phi} is the value of \code{x} at
which the curve is equal to its mid-point, i.e. 1 / 2.
}
