% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{fit.logistic4}
\alias{fit.logistic4}
\alias{fit_constrained.logistic4}
\title{4-parameter logistic fit}
\usage{
\method{fit}{logistic4}(object)

\method{fit_constrained}{logistic4}(object)
}
\arguments{
\item{object}{object of class \code{logistic4}.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value found of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Fit a 4-parameter logistic function to observed data with a Maximum
Likelihood approach.
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined here as

\code{alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate (also known as the Hill coefficient),
and \code{phi} is the value of \code{x} at which the curve is equal to its mid-point.
}
