% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{fit.logistic5}
\alias{fit.logistic5}
\alias{fit_constrained.logistic5}
\title{5-parameter logistic fit}
\usage{
\method{fit}{logistic5}(object)

\method{fit_constrained}{logistic5}(object)
}
\arguments{
\item{object}{object of class \code{logistic5}.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value found of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Fit a 5-parameter logistic function to observed data with a Maximum
Likelihood approach.
}
\details{
The 5-parameter logistic function \verb{f(x; theta)} is defined here as

\code{alpha + (beta - alpha) / (1 + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu)}, \code{beta > alpha}, and \code{nu > 0}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\code{alpha + (beta - alpha) / (1 + exp(v - eta * (x - phi)))^(-exp(-v))}
}
