% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fit_nls}
\alias{fit_nls}
\title{Algorithm initialization}
\usage{
fit_nls(object, rss_fn, formula, start)
}
\arguments{
\item{object}{object of some model class.}

\item{rss_fn}{residual sum of squares to be minimized.}

\item{formula}{a nonlinear model formula including variables and parameters.}

\item{start}{matrix of candidate starting points.}
}
\description{
Use \code{nls} to find a good initial value.
}
