% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fit_optim}
\alias{fit_optim}
\title{Algorithm initialization}
\usage{
fit_optim(object, fn, start)
}
\arguments{
\item{object}{object of some model class.}

\item{fn}{residual sum of squares to be minimized.}

\item{start}{candidate starting point.}
}
\description{
Use \code{optim} to find a good initial value.
}
