% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{fn.logistic4}
\alias{fn.logistic4}
\alias{fn.logistic4_fit}
\title{4-parameter logistic function}
\usage{
\method{fn}{logistic4}(object, x, theta)

\method{fn}{logistic4_fit}(object, x, theta)
}
\arguments{
\item{object}{object of class \code{logistic4}.}

\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{theta}{numeric vector with the four parameters in the form
\code{c(alpha, omega, eta, phi)}.}
}
\value{
Numeric vector with the values of the logistic function.
}
\description{
Evaluate at a particular set of parameters the 4-parameter logistic function.
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + omega / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, omega, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{omega} is the width of the curve, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.
}
