% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{fn.logistic5}
\alias{fn.logistic5}
\alias{fn.logistic5_fit}
\title{5-parameter logistic function}
\usage{
\method{fn}{logistic5}(object, x, theta)

\method{fn}{logistic5_fit}(object, x, theta)
}
\arguments{
\item{object}{object of class \code{logistic5}.}

\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{theta}{numeric vector with the five parameters in the form
\code{c(alpha, beta, eta, phi, nu)}.}
}
\value{
Numeric vector with the values of the logistic function.
}
\description{
Evaluate at a particular set of parameters the 5-parameter logistic function.
}
\details{
The 5-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (1 + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu)}.
}
