% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{gradient_hessian.gompertz}
\alias{gradient_hessian.gompertz}
\title{Gompertz function}
\usage{
\method{gradient_hessian}{gompertz}(object, theta)
}
\arguments{
\item{object}{object of class \code{gompertz}.}

\item{theta}{numeric vector with the four parameters in the form
\code{c(alpha, beta, eta, phi)}.}
}
\value{
List of two elements: \code{G} the gradient and \code{H} the Hessian.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
Gompertz function.
}
\details{
The Gompertz function \verb{f(x; theta)} is defined here as

\verb{alpha + (beta - alpha) exp(-exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate, and \code{phi} related to the function
value at \code{x = 0}.
}
