% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{gradient_hessian.logistic2}
\alias{gradient_hessian.logistic2}
\title{2-parameter logistic function}
\usage{
\method{gradient_hessian}{logistic2}(object, theta)
}
\arguments{
\item{object}{object of class \code{logistic2}.}

\item{theta}{numeric vector with the parameters in the form \code{c(eta, phi)}.}
}
\value{
List of two elements: \code{G} the gradient and \code{H} the Hessian.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
2-parameter logistic function.
}
\details{
The 2-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{1 / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(eta, phi)}, \code{eta} is the steepness of the curve or growth
rate (also known as the Hill coefficient), and \code{phi} is the value of \code{x} at
which the curve is equal to its mid-point, i.e. 1 / 2.
}
