% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{init.gompertz}
\alias{init.gompertz}
\title{Initialize vector of parameters}
\usage{
\method{init}{gompertz}(object)
}
\arguments{
\item{object}{object of class \code{gompertz}.}
}
\value{
Numeric vector of length 4 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the four parameters of the Gompertz function.
}
