% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{init.logistic2}
\alias{init.logistic2}
\title{Initialize vector of parameters}
\usage{
\method{init}{logistic2}(object)
}
\arguments{
\item{object}{object of class \code{logistic2}.}
}
\value{
Numeric vector of length 2 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the four parameters of the logistic function.
}
