% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{init.logistic4}
\alias{init.logistic4}
\title{Initialize vector of parameters}
\usage{
\method{init}{logistic4}(object)
}
\arguments{
\item{object}{object of class \code{logistic4}.}
}
\value{
Numeric vector of length 4 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the four parameters of the logistic function.
}
