% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{init.logistic5}
\alias{init.logistic5}
\title{Initialize vector of parameters}
\usage{
\method{init}{logistic5}(object)
}
\arguments{
\item{object}{object of class \code{logistic5}.}
}
\value{
Numeric vector of length 5 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the five parameters of the logistic function.
}
