% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{init.logistic6}
\alias{init.logistic6}
\title{Initialize vector of parameters}
\usage{
\method{init}{logistic6}(object)
}
\arguments{
\item{object}{object of class \code{logistic6}.}
}
\value{
Numeric vector of length 6 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the six parameters of the logistic function.
}
