% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{logistic2_fisher_info_normal}
\alias{logistic2_fisher_info_normal}
\title{2-parameter logistic fit}
\usage{
logistic2_fisher_info_normal(stats, n, theta, sigma)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}

\item{n}{effective sample size.}

\item{theta}{numeric vector with the model parameters.}

\item{sigma}{estimate of the standard deviation.}
}
\value{
Fisher information matrix evaluated at \code{theta}.
}
\description{
Evaluate the Fisher information matrix at the maximum likelihood estimate.
}
\details{
Let \verb{mu(x; theta)} be the 2-parameter logistic function. We assume that our
observations \code{y} are independent and such that
\verb{y = mu(x; theta) + sigma * epsilon}, where \code{epsilon} has a standard Normal
distribution \code{N(0, 1)}.

The 2-by-2 (symmetric) Fisher information matrix is the expected value of
the negative Hessian matrix of the log-likelihood function.
}
