% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{logistic2_fit}
\alias{logistic2_fit}
\alias{logistic2_fit_constrained}
\alias{logistic2_fit_weighted}
\alias{logistic2_fit_weighted_constrained}
\title{2-parameter logistic fit}
\usage{
logistic2_fit(x, y, start, max_iter)

logistic2_fit_constrained(x, y, start, max_iter, lower_bound, upper_bound)

logistic2_fit_weighted(x, y, w, start, max_iter)

logistic2_fit_weighted_constrained(
  x,
  y,
  w,
  start,
  max_iter,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{lower_bound}{numeric vector of length 2 with the minimum admissible
value of \code{eta} and \code{phi} respectively. Values can be equal to \code{-Inf}.}

\item{upper_bound}{numeric vector of length 4 with the maximum admissible
value of \code{eta} and \code{phi} respectively. Values can be equal to \code{Inf}.}

\item{w}{numeric vector of optional weights.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value found of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Fit a 2-parameter logistic function to observed data with a Maximum
Likelihood approach.
}
\details{
The 2-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{1 / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(eta, phi)}, \code{eta} is the steepness of the curve or growth
rate (also known as the Hill coefficient), and \code{phi} is the value of \code{x} at
which the curve is equal to its mid-point, i.e. 1 / 2.
}
