% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{logistic2_gradient_hessian}
\alias{logistic2_gradient_hessian}
\title{2-parameter logistic function}
\usage{
logistic2_gradient_hessian(x, theta)
}
\arguments{
\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{theta}{numeric vector with the four parameters in the form
\code{c(eta, phi)}.}
}
\value{
List of two elements. Element \code{G} is a numeric matrix of dimension
length(x)-by-2, where each row is the gradient of the logistic function at
the corresponding element of \code{x}. Element \code{H} is an array of dimension
length(x)-by-2-by-2, where \code{H[k, , ]} is the 2-by-2 Hessian matrix at
\code{x[k]}.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
2-parameter logistic function.
}
\details{
The 2-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{1 / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(eta, phi)}, \code{eta} is the steepness of the curve or growth
rate (also known as the Hill coefficient), and \code{phi} is the value of \code{x} at
which the curve is equal to its mid-point, i.e. 1 / 2.
}
