% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{logistic2_init}
\alias{logistic2_init}
\title{Initialize vector of parameters}
\usage{
logistic2_init(stats)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}
}
\value{
Numeric vector of length 2 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the four parameters of the logistic function.
}
