% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{logistic2_rss}
\alias{logistic2_rss}
\alias{logistic2_rss_fixed}
\title{Residual sum of squares}
\usage{
logistic2_rss(stats)

logistic2_rss_fixed(stats, known_param)
}
\arguments{
\item{stats}{matrix of sufficient statistics.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any.}
}
\value{
Function handle \code{f(theta)} to evaluate the RSS associated to a
particular parameter choice \code{theta}.
}
\description{
Evaluate the residual sum of squares (RSS) against the mean of a
2-parameter logistic model.
}
\details{
The 2-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{1 / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(eta, phi)}, \code{eta} is the steepness of the curve or growth
rate (also known as the Hill coefficient), and \code{phi} is the value of \code{x} at
which the curve is equal to its mid-point, i.e. 1 / 2.
}
