% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_common.R
\name{logistic4_function}
\alias{logistic4_function}
\title{4-parameter logistic function}
\usage{
logistic4_function(x, param)
}
\arguments{
\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{param}{numeric vector with the four parameters in the form
\code{c(alpha, beta, eta, phi)}.}
}
\value{
Numeric vector of the same length of \code{x} with the values of the
logistic function.
}
\description{
Evaluate at a particular set of parameters the 4-parameter logistic function.
}
\details{
The 4-parameter logistic function is defined in this package as

\verb{f(x; theta) = alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta} is the upper horizontal asymptote, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.
}
