% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{logistic4_gradient_hessian}
\alias{logistic4_gradient_hessian}
\title{4-parameter logistic function}
\usage{
logistic4_gradient_hessian(x, theta)
}
\arguments{
\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{theta}{numeric vector with the four parameters in the form
\code{c(alpha, beta, eta, phi)}.}
}
\value{
List of two elements. Element \code{G} is a numeric matrix of dimension
length(x)-by-4, where each row is the gradient of the logistic function at
the corresponding element of \code{x}. Element \code{H} is an array of dimension
length(x)-by-4-by-4, where \code{H[k, , ]} is the 4-by-4 Hessian matrix at
\code{x[k]}.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
4-parameter logistic function.
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta} is the upper horizontal asymptote, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\code{alpha + exp(omega) / (1 + exp(-eta * (x - phi)))}

where \code{omega} is the log-width of the curve. Note that it is simply
\code{beta = alpha + exp(omega)}.

NOTE: This function expects the input vector
\code{theta = c(alpha, beta, eta, phi)} but computes the gradient and Hessian of
the alternative parametrization \verb{theta' = (alpha, omega, eta, phi)}.
This duality is unfortunately a trade-off between what users expect (first
version is easier to interpret) and the optimization stability of the second
version.
}
