% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_common.R
\name{logistic4_init}
\alias{logistic4_init}
\title{Initialize vector of parameters}
\usage{
logistic4_init(stats)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}
}
\value{
Numeric vector of length 4 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the four parameters of the logistic function.
}
