% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_common.R
\name{logistic4_loglik}
\alias{logistic4_loglik}
\title{Log-likelihood value}
\usage{
logistic4_loglik(variance, n, k = 4, log_w = 0)
}
\arguments{
\item{variance}{corrected maximum likelihood estimate of the variance.}

\item{n}{total sample size.}

\item{k}{number of estimated parameters, with a default of 4 in case of
unconstrained optimization.}

\item{log_w}{sum of log-weights, with a default value of 0 in case of weights
all equal to 1.}
}
\value{
Corrected maximum likelihood estimator of the variance.
}
\description{
Evaluate the maximum of the log-likelihood function.
}
