% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{logistic4_ntrm}
\alias{logistic4_ntrm}
\alias{logistic4_ntrm_constrained}
\title{Fit a 4-parameter logistic function}
\usage{
logistic4_ntrm(stats, start, max_iter)

logistic4_ntrm_constrained(
  stats,
  start,
  max_iter,
  constraint,
  lower_bound,
  upper_bound,
  known_param
)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{constraint}{boolean matrix with a representation of the constraints.}

\item{lower_bound}{numeric vector of length 4 with the minimum admissible
values.}

\item{upper_bound}{numeric vector of length 4 with the maximum admissible
values.}

\item{known_param}{numeric vector of fixed known parameters.}
}
\value{
A list with the following components:
\describe{
\item{optimum}{maximum likelihood estimates of the model parameters.}
\item{minimum}{(local) minimum of the residual sum of squares around the
means.}
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
}
}
\description{
Use a Newton trust-region method to fit a 4-parameter logistic function to
observed data.
}
