% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_fit.R
\name{logistic4_ntrm_constrained}
\alias{logistic4_ntrm_constrained}
\alias{logistic4_ntrm_unconstrained}
\title{Fit a four parameter logistic function}
\usage{
logistic4_ntrm_constrained(
  stats,
  init,
  max_iter,
  constraint,
  lower_bound,
  upper_bound,
  known_param
)

logistic4_ntrm_unconstrained(stats, init, max_iter)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}

\item{init}{numeric vector of length 4 with an initial guess of the solution.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{constraint}{boolean matrix with a representation of the constraints.}

\item{lower_bound}{numeric vector of length 4 with the minimum admissible
values.}

\item{upper_bound}{numeric vector of length 4 with the maximum admissible
values.}

\item{known_param}{numeric vector of fixed known parameters.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean value assessing if the optimization algorithm
converged or not.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean value set to \code{TRUE} if optimization was
constrained.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{sigma}{corrected maximum likelihood estimate of the standard
deviation.}
\item{rss}{value of the residual sum of squares.}
\item{loglik}{maximum value of the log-likelihood function.}
\item{df.residual}{residual degrees of freedom.}
}
}
\description{
Use a Newton trust-region method to fit a four parameter logistic function to
observed data.
}
