% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_rss.R
\name{logistic4_rss_hessian}
\alias{logistic4_rss_hessian}
\alias{logistic4_rss_hessian_fixed}
\title{Residual sum of squares}
\usage{
logistic4_rss_hessian(stats)

logistic4_rss_hessian_fixed(stats, known_param)
}
\arguments{
\item{stats}{matrix of sufficient statistics. First column are the unique
elements of \code{x}, second column are the corresponding sample sizes,
third column are the sample means, while the fourth column are the
(uncorrected) sample variances.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any. Vector is of the form \code{c(alpha, beta, eta, phi)}
where \code{NA} represents an unknown parameter.}
}
\value{
Function handle \code{f(theta)} to evaluate the Hessian matrix of the RSS
associated to a particular parameter choice \code{theta}.
}
\description{
Evaluate the Hessian matrix of the residual sum of squares (RSS) of a
4-parameter logistic model.
}
\details{
The 4-parameter logistic function is defined in this package as

\verb{f(x; theta) = alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta} is the upper horizontal asymptote, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.
}
