% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_common.R
\name{logistic4_suff_stat}
\alias{logistic4_suff_stat}
\title{Sufficient statistics}
\usage{
logistic4_suff_stat(x, y)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}
}
\value{
Numeric matrix where the first column \code{x} are the sorted unique
values of input vector \code{x}, second column \code{n} are the corresponding number
of observations, third column \code{m} are the sample means, and fourth column
\code{v} are the (uncorrected) sample variances.
}
\description{
Compute sufficient statistics for each unique value of the predictor
variable.
}
