% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_common.R
\name{logistic4_variance}
\alias{logistic4_variance}
\title{Variance estimator}
\usage{
logistic4_variance(rss, n, k = 4)
}
\arguments{
\item{rss}{value of the residual sum of squares, that is its minimum possible
value because we are using maximum likelihood estimation.}

\item{n}{total sample size.}

\item{k}{number of estimated parameters, with a default of 4 in case of
unconstrained optimization.}
}
\value{
Corrected maximum likelihood estimate of the variance.
}
\description{
Compute the corrected variance estimate associated to a 4-parameter logistic
function.
}
