% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_weighted_fit.R
\name{logistic4_weighted_fisher_info}
\alias{logistic4_weighted_fisher_info}
\title{4-parameter logistic fit}
\usage{
logistic4_weighted_fisher_info(x, w, param, sigma)
}
\arguments{
\item{x}{numeric vector at which the function is to be evaluated.}

\item{w}{numeric vector of observation weights.}

\item{param}{numeric vector with the four parameters in the form
\code{c(alpha, beta, eta, phi)}.}

\item{sigma}{estimate of the standard deviation.}
}
\value{
Fisher information matrix evaluated at \code{param}.
}
\description{
Evaluate the Fisher information matrix at the maximum likelihood estimate.
}
\details{
The 4-parameter logistic function is defined in this package as

\verb{f(x; theta) = alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta} is the upper horizontal asymptote, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point. The model
is extended by also assuming that \verb{y = f(x; theta) + sigma * z / w}, where
\code{z} is a standard normal random variable. The 5-by-5 (symmetric) Fisher
information matrix is the expected value of the negative Hessian matrix of
the log-likelihood function.
}
