% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_weighted_fit.R
\name{logistic4_weighted_fit_constrained}
\alias{logistic4_weighted_fit_constrained}
\alias{logistic4_weighted_fit_unconstrained}
\title{4-parameter logistic fit}
\usage{
logistic4_weighted_fit_constrained(
  x,
  y,
  w,
  start,
  max_iter,
  lower_bound,
  upper_bound
)

logistic4_weighted_fit_unconstrained(x, y, w, start, max_iter)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{w}{numeric vector of observation weights.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{lower_bound}{numeric vector of length 4 with the minimum admissible
value of \code{alpha}, \code{beta}, \code{eta}, and \code{phi} respectively. Values can be
equal to \code{-Inf}.}

\item{upper_bound}{numeric vector of length 4 with the maximum admissible
value of \code{alpha}, \code{beta}, \code{eta}, and \code{phi} respectively. Values can be
equal to \code{Inf}.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean value assessing if the optimization algorithm
converged or not.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean value set to \code{TRUE} if optimization was
constrained.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{sigma}{corrected maximum likelihood estimate of the standard
deviation.}
\item{rss}{value of the residual sum of squares.}
\item{loglik}{maximum value of the log-likelihood function.}
\item{df.residual}{residual degrees of freedom.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{fitted.values}{the fitted mean values.}
\item{residuals}{the residuals, that is response minus fitted values.}
\item{weights}{the specified weights.}
\item{fisher_info}{Fisher information matrix evaluated at the maximum
likelihood estimator.}
}
}
\description{
Fit a 4-parameter logistic function to observed data with a Maximum
Likelihood approach.
}
\details{
The 4-parameter logistic function is defined in this package as

\verb{f(x; theta) = alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta} is the upper horizontal asymptote, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.

The model assumes the observed data to be normally distributed around the
logistic function, with a possibly unequal standard deviation at each and
every predictor \code{x[k]}, i.e.

\verb{y[i] ~ N(f(x[k]; theta), sigma^2 / w[i])}

for any predictor \code{x[k]} and for any statistical unit \code{i} observed at
\code{x[k]}.
}
