% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4_weighted_rss.R
\name{logistic4_weighted_rss_gradient}
\alias{logistic4_weighted_rss_gradient}
\alias{logistic4_weighted_rss_gradient_fixed}
\title{Residual sum of squares}
\usage{
logistic4_weighted_rss_gradient(x, y, w)

logistic4_weighted_rss_gradient_fixed(x, y, w, known_param)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{w}{numeric vector of observation weights.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any. Vector is of the form \code{c(alpha, beta, eta, phi)}
where \code{NA} represents an unknown parameter.}
}
\value{
Function handle \code{f(theta)} to evaluate the gradient of the RSS
associated to a particular parameter choice \code{theta}.
}
\description{
Evaluate the gradient of the residual sum of squares (RSS) of a 4-parameter
logistic model when each observation is given its own weight.
}
\details{
The 4-parameter logistic function is defined in this package as

\verb{f(x; theta) = alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta} is the upper horizontal asymptote, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.
}
