% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{logistic5_gradient_hessian}
\alias{logistic5_gradient_hessian}
\title{5-parameter logistic function}
\usage{
logistic5_gradient_hessian(x, theta)
}
\arguments{
\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{theta}{numeric vector with the five parameters in the form
\code{c(alpha, beta, eta, phi, nu)}.}
}
\value{
List of two elements. Element \code{G} is a numeric matrix of dimension
length(x)-by-5, where each row is the gradient of the logistic function at
the corresponding element of \code{x}. Element \code{H} is an array of dimension
length(x)-by-5-by-5, where \code{H[k, , ]} is the 5-by-5 Hessian matrix at
\code{x[k]}.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
5-parameter logistic function.
}
\details{
The 5-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (1 + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu)}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega) / (1 + exp(nu' - eta * (x - phi)))^(-exp(-nu'))}

Note that it is simply \code{beta = alpha + exp(omega)} and \verb{nu = exp(nu')}.

NOTE: This function expects the input vector
\code{theta = c(alpha, beta, eta, phi, nu)} but computes the gradient and
Hessian of the alternative parametrization
\verb{theta' = (alpha, omega, eta, phi, nu')}.
This duality is unfortunately a trade-off between what users expect (first
version is easier to interpret) and the optimization stability of the second
version.
}
