% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{logistic5_init}
\alias{logistic5_init}
\title{Initialize vector of parameters}
\usage{
logistic5_init(stats)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}
}
\value{
Numeric vector of length 5 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the five parameters of the logistic function.
}
