% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{logistic6_fit}
\alias{logistic6_fit}
\alias{logistic6_fit_constrained}
\alias{logistic6_fit_weighted}
\alias{logistic6_fit_weighted_constrained}
\title{6-parameter logistic fit}
\usage{
logistic6_fit(x, y, start, max_iter)

logistic6_fit_constrained(x, y, start, max_iter, lower_bound, upper_bound)

logistic6_fit_weighted(x, y, w, start, max_iter)

logistic6_fit_weighted_constrained(
  x,
  y,
  w,
  start,
  max_iter,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{lower_bound}{numeric vector of length 6 with the minimum admissible
value of \code{alpha}, \code{omega}, \code{eta}, \code{phi}, \verb{nu'}, and \verb{xi'} respectively.
Values can be equal to \code{-Inf}.}

\item{upper_bound}{numeric vector of length 6 with the maximum admissible
value of \code{alpha}, \code{omega}, \code{eta}, \code{phi}, \verb{nu'}, and \verb{xi'} respectively.
Values can be equal to \code{Inf}.}

\item{w}{numeric vector of optional weights.}
}
\value{
A list with the following components:
\describe{
\item{converged}{boolean. \code{TRUE} if the optimization algorithm converged,
\code{FALSE} otherwise.}
\item{iterations}{total number of iterations performed by the
optimization algorithm}
\item{constrained}{boolean. \code{TRUE} if optimization was constrained,
\code{FALSE} otherwise.}
\item{estimated}{boolean vector indicating which parameters were
estimated from the data.}
\item{coefficients}{maximum likelihood estimates of the model
parameters.}
\item{rss}{minimum value found of the residual sum of squares.}
\item{df.residual}{residual degrees of freedom.}
\item{fitted.values}{fitted mean values.}
\item{residuals}{residuals, that is response minus fitted values.}
\item{weights}{vector of weights used for the fit.}
}
}
\description{
Fit a 6-parameter logistic function to observed data with a Maximum
Likelihood approach.
}
\details{
The 6-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (xi + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu, xi)}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega) / (exp(xi') + exp(nu' - eta * (x - phi)))^(-exp(-nu'))}

Note that it is simply \code{beta = alpha + exp(omega)}, \verb{nu = exp(nu')},
\verb{xi = exp(xi')}.
}
