% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{logistic6_gradient_hessian}
\alias{logistic6_gradient_hessian}
\title{6-parameter logistic function}
\usage{
logistic6_gradient_hessian(x, theta)
}
\arguments{
\item{x}{numeric vector at which the logistic function is to be evaluated.}

\item{theta}{numeric vector with the six parameters in the form
\code{c(alpha, beta, eta, phi, nu, xi)}.}
}
\value{
List of two elements. Element \code{G} is a numeric matrix of dimension
length(x)-by-6, where each row is the gradient of the logistic function at
the corresponding element of \code{x}. Element \code{H} is an array of dimension
length(x)-by-6-by-6, where \code{H[k, , ]} is the 6-by-6 Hessian matrix at
\code{x[k]}.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
6-parameter logistic function.
}
\details{
The 6-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (xi + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu, xi)}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega) / (exp(xi') + exp(nu' - eta * (x - phi)))^(-exp(-nu'))}

Note that it is simply \code{beta = alpha + exp(omega)}, \verb{nu = exp(nu')},
\verb{xi = exp(xi')}.

NOTE: This function expects the input vector
\code{theta = c(alpha, beta, eta, phi, nu, xi)} but computes the gradient and
Hessian of the alternative parametrization
\verb{theta' = (alpha, omega, eta, phi, nu', xi')}.
This duality is unfortunately a trade-off between what users expect (first
version is easier to interpret) and the optimization stability of the second
version.
}
