% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{logistic6_init}
\alias{logistic6_init}
\title{Initialize vector of parameters}
\usage{
logistic6_init(stats)
}
\arguments{
\item{stats}{numeric matrix of sufficient statistics.}
}
\value{
Numeric vector of length 6 with a (hopefully) good starting point.
}
\description{
Given the sufficient statistics, try to guess a good approximation to the
Maximum Likelihood estimator of the six parameters of the logistic function.
}
