% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R, R/logistic2.R, R/logistic4.R,
%   R/logistic5.R, R/logistic6.R
\name{gompertz_new}
\alias{gompertz_new}
\alias{logistic2_new}
\alias{logistic4_new}
\alias{logistic5_new}
\alias{logistic6_new}
\title{Fit dose-response data}
\usage{
gompertz_new(x, y, w, start, max_iter, lower_bound, upper_bound)

logistic2_new(x, y, w, start, max_iter, lower_bound, upper_bound)

logistic4_new(x, y, w, start, max_iter, lower_bound, upper_bound)

logistic5_new(x, y, w, start, max_iter, lower_bound, upper_bound)

logistic6_new(x, y, w, start, max_iter, lower_bound, upper_bound)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{w}{an optional vector of weights to be used in the fitting
process.}

\item{start}{starting values for the parameters.}

\item{max_iter}{maximum number of iterations in the optimization algorithm.}

\item{lower_bound}{numeric vector with the minimum admissible values of the
parameters.}

\item{upper_bound}{numeric vector with the maximum admissible values of the
parameters.}
}
\value{
An object of class \verb{logistic*}.
}
\description{
Use a Newton trust-region method to fit a logistic function to dose-response
data.
}
