% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{logistic6_rss}
\alias{logistic6_rss}
\alias{logistic6_rss_fixed}
\title{Residual sum of squares}
\usage{
logistic6_rss(stats)

logistic6_rss_fixed(stats, known_param)
}
\arguments{
\item{stats}{matrix of sufficient statistics.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any.}
}
\value{
Function handle \code{f(theta)} to evaluate the RSS associated to a
particular parameter choice \code{theta}.
}
\description{
Evaluate the residual sum of squares (RSS) against the mean of a
6-parameter logistic model.
}
\details{
The 6-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (xi + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu, xi)}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega) / (exp(xi') + exp(nu' - eta * (x - phi)))^(-exp(-nu'))}

Note that it is simply \code{beta = alpha + exp(omega)}, \verb{nu = exp(nu')},
\verb{xi = exp(xi')}.
}
