% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{loglik_normal}
\alias{loglik_normal}
\title{Log-likelihood value}
\usage{
loglik_normal(deviance, n, log_w = 0)
}
\arguments{
\item{deviance}{deviance value, i.e. residual sum of squares.}

\item{n}{total sample size.}

\item{log_w}{sum of log-weights.}
}
\value{
Value of the log-likelihood function at the MLE.
}
\description{
Evaluate the maximum of the log-likelihood function associated with a Normal
distribution.
}
