% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogistic2.R
\name{loglogistic2_gradient_2}
\alias{loglogistic2_gradient_2}
\alias{loglogistic2_hessian_2}
\alias{loglogistic2_gradient_hessian_2}
\title{2-parameter log-logistic function gradient and Hessian}
\usage{
loglogistic2_gradient_2(x, theta, delta)

loglogistic2_hessian_2(x, theta, delta)

loglogistic2_gradient_hessian_2(x, theta, delta)
}
\arguments{
\item{x}{numeric vector at which the function is to be evaluated.}

\item{theta}{numeric vector with the two parameters in the form
\code{c(eta, phi)}.}

\item{delta}{value of delta parameter (either 1 or -1).}
}
\value{
Gradient or Hessian of the alternative parameterization evaluated at
the specified point.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
2-parameter log-logistic function.
}
\details{
The 2-parameter log-logistic function \verb{f(x; theta)} is defined here as

\verb{g(x; theta) = x^eta / (x^eta + phi^eta)}
\verb{f(x; theta) = alpha + delta g(x; theta)}

where \code{x >= 0}, \code{theta = c(alpha, delta, eta, phi)}, \code{eta > 0}, and
\code{phi > 0}. Only \code{eta} and \code{phi} are free to vary (therefore the name), while
\code{c(alpha, delta)} are constrained to be either \code{c(0, 1)} (monotonically
increasing curve) or \code{c(1, -1)} (monotonically decreasing curve).

This set of functions use a different parameterization from
\code{link[drda]{loglogistic2_gradient}}. To avoid the non-negative
constraints of parameters, the gradient and Hessian computed here are for
the function with \code{eta2 = log(eta)} and \code{phi2 = log(phi)}.

Note that argument \code{theta} is on the original scale and not on the log scale.
}
