% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{mle_asy.gompertz}
\alias{mle_asy.gompertz}
\title{Maximum likelihood estimators}
\usage{
\method{mle_asy}{gompertz}(object, theta)
}
\arguments{
\item{object}{object of class \code{gompertz}.}

\item{theta}{vector of parameters.}
}
\value{
Numeric vector of length 2 with the MLE of the two asymptotes.
}
\description{
Given a set of parameters, compute the maximum likelihood estimates of the
lower and upper horizontal asymptotes.
}
