% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{mle_asy.logistic2}
\alias{mle_asy.logistic2}
\title{Maximum likelihood estimators}
\usage{
\method{mle_asy}{logistic2}(object, theta)
}
\arguments{
\item{object}{object of class \code{logistic2}.}

\item{theta}{vector of parameters.}
}
\value{
Numeric vector \code{theta}.
}
\description{
Given a set of parameters, compute the maximum likelihood estimates of the
lower and upper horizontal asymptotes.
}
