% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{nauc.gompertz_fit}
\alias{nauc.gompertz_fit}
\alias{naac.gompertz_fit}
\title{Gompertz fit}
\usage{
\method{nauc}{gompertz_fit}(object, lower_bound = -10, upper_bound = 10)

\method{naac}{gompertz_fit}(object, lower_bound = -10, upper_bound = 10)
}
\arguments{
\item{object}{object of class \code{gompertz_fit}.}

\item{lower_bound}{numeric value with the lower bound of the integration
interval.}

\item{upper_bound}{numeric value with the upper bound of the integration
interval.}
}
\value{
Numeric value with the requested area.
}
\description{
Evaluate the normalized area under the curve (AUC) and area above the curve
(AAC).
}
\details{
The Gompertz function \verb{f(x; theta)} is defined here as

\verb{alpha + (beta - alpha) exp(-exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate, and \code{phi} related to the function
value at \code{x = 0}.

The area under the curve (AUC) is simply the integral of \verb{f(x; theta)}
between \code{lower_bound} and \code{upper_bound} with respect to \code{x}.

When the interval of integration is fixed, since the the curve ranges between
\code{alpha} and \code{beta}, the curve \verb{f(x; theta)} is contained into the rectangle
of height \code{beta - alpha} and width \code{upper_bound - lower_bound}. The maximum
area the curve can have is obviously
\code{(upper_bound - lower_bound) * (beta - alpha)}.

We first shift the curve by \code{alpha} to set the minimum to 0. We then
integrate the curve and define the normalized AUC (NAUC) by dividing its
value by the maximum area. As a consequence, the normalized area above the
curve is simply \code{NAAC = 1 - NAUC}.

Default values of \code{lower_bound} and \code{upper_bound} were chosen based on common
dose ranges used in the literature. They are also symmetric around zero
so that \code{NAUC} and \code{NAAC} are equal to \code{0.5} in the standard logistic model.
}
