% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic2.R
\name{nauc.logistic2_fit}
\alias{nauc.logistic2_fit}
\alias{naac.logistic2_fit}
\title{2-parameter logistic fit}
\usage{
\method{nauc}{logistic2_fit}(object, lower_bound = -10, upper_bound = 10)

\method{naac}{logistic2_fit}(object, lower_bound = -10, upper_bound = 10)
}
\arguments{
\item{object}{object of class \code{logistic2_fit}.}

\item{lower_bound}{numeric value with the lower bound of the integration
interval.}

\item{upper_bound}{numeric value with the upper bound of the integration
interval.}
}
\value{
Numeric value with the requested area.
}
\description{
Evaluate the normalized area under the curve (AUC) and area above the curve
(AAC).
}
\details{
The 2-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{1 / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(eta, phi)}, \code{eta} is the steepness of the curve or growth
rate (also known as the Hill coefficient), and \code{phi} is the value of \code{x} at
which the curve is equal to its mid-point, i.e. 1 / 2.

The area under the curve (AUC) is simply the integral of \verb{f(x; theta)}
between \code{lower_bound} and \code{upper_bound} with respect to \code{x}.

When the interval of integration is fixed, since the maximum value the curve
can attain is 1, the curve \verb{f(x; theta)} is contained into the rectangle
of height 1 and width \code{upper_bound - lower_bound}. The maximum area the curve
can have is obviously \code{upper_bound - lower_bound}.

The normalized AUC is simply \code{NAUC = AUC / (upper_bound - lower_bound)}. As a
consequence, the normalized area above the curve is \code{NAAC = 1 - NAUC}.

Default values of \code{lower_bound} and \code{upper_bound} were chosen based on common
dose ranges used in the literature. They are also symmetric around zero
so that \code{NAUC} and \code{NAAC} are equal to \code{0.5} in the standard logistic model.
}
