% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{nauc.logistic4_fit}
\alias{nauc.logistic4_fit}
\alias{naac.logistic4_fit}
\title{4-parameter logistic fit}
\usage{
\method{nauc}{logistic4_fit}(object, lower_bound = -10, upper_bound = 10)

\method{naac}{logistic4_fit}(object, lower_bound = -10, upper_bound = 10)
}
\arguments{
\item{object}{object of class \code{logistic4_fit}.}

\item{lower_bound}{numeric value with the lower bound of the integration
interval.}

\item{upper_bound}{numeric value with the upper bound of the integration
interval.}
}
\value{
Numeric value with the requested area.
}
\description{
Evaluate the normalized area under the curve (AUC) and area above the curve
(AAC).
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + omega / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, omega, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{omega} is the width of the curve, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.

The area under the curve (AUC) is simply the integral of \verb{f(x; theta)}
between \code{lower_bound} and \code{upper_bound} with respect to \code{x}.

When the interval of integration is fixed, since the the curve ranges between
\code{alpha} and \code{alpha + omega}, the curve \verb{f(x; theta)} is contained into the
rectangle of height \code{omega} and width \code{upper_bound - lower_bound}. The
maximum area the curve can have is obviously
\code{(upper_bound - lower_bound) * omega}.

We first shift the curve by \code{alpha} to set the minimum to 0. We then
integrate the curve and define the normalized AUC (NAUC) by dividing its
value by the maximum area. As a consequence, the normalized area above the
curve is simply \code{NAAC = 1 - NAUC}.

Default values of \code{lower_bound} and \code{upper_bound} were chosen based on common
dose ranges used in the literature. They are also symmetric around zero
so that \code{NAUC} and \code{NAAC} are equal to \code{0.5} in the standard logistic model.
}
