% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{nauc}
\alias{nauc}
\title{Area under the curve}
\usage{
nauc(object, lower_bound = -10, upper_bound = 10)
}
\arguments{
\item{object}{fit object as returned by \code{\link[drda]{drda}}.}

\item{lower_bound}{numeric value with the lower bound of the integration
interval.}

\item{upper_bound}{numeric value with the upper bound of the integration
interval.}
}
\value{
Numeric value representing the normalized area under the curve.
}
\description{
Evaluate the normalized area under the curve (AUC).
}
\details{
The area under the curve (AUC) is simply the integral of the chosen model
\verb{f(x; theta)} between \code{lower_bound} and \code{upper_bound} with respect to \code{x}.

When the interval of integration is fixed, the curve \verb{f(x; theta)} is
contained into the rectangle of height \code{omega} and width
\code{upper_bound - lower_bound}. The maximum area the curve can have is obviously
\code{(upper_bound - lower_bound) * omega}.

We first shift the curve to set its minimum to 0. We then integrate the curve
and define the normalized AUC (NAUC) by dividing its value by the maximum
area.

Default values of \code{lower_bound} and \code{upper_bound} were chosen based on common
dose ranges used in the literature. They are also symmetric around zero
so that \code{NAUC} is equal to \code{0.5} in the standard logistic model.
}
