% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm.R
\name{ntrm}
\alias{ntrm}
\title{Newton with Trust Region Method}
\usage{
ntrm(fn, gh, init, max_iter, update_fn = NULL)
}
\arguments{
\item{fn}{function handle to evaluate the function to minimize.}

\item{gh}{function handle to evaluate both the gradient and Hessian of the
function to minimize.}

\item{init}{numeric vector of starting values.}

\item{max_iter}{integer value for the maximum number of iterations.}

\item{update_fn}{function handle to update a subset of parameters with the
actual optimum.}
}
\value{
A list containing the value \code{optimum} for which \code{f(optimum)} is
minimum, the value \code{minimum} containing \code{f(optimum)}, and a boolean
value \code{converged} stating if the algorithm converged or not.
}
\description{
Find the minimum of a function using the Newton's method, combined with the
Trust Region method.
}
\details{
The code is adapted from the Julia implementation of
\href{https://github.com/JuliaNLSolvers/Optim.jl}{Optim.jl}. See
'src/multivariate/solvers/second_order/newton_trust_region.jl'
file for the original source code.
}
\references{
Jorge J Moré and D C Sorensen. Computing a Trust Region Step.
\strong{SIAM Journal on Scientific and Statistical Computing}, 4(3):553-572, 1983.
doi: 10.1137/0904038.

Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.

Patrick Kofod Mogensen and Asbjørn Nilsen Riseth.
Optim: A mathematical optimization package for Julia.
\strong{Journal of Open Source Software}, 3(24):615, 2018.
doi: 10.21105/joss.00615.
}
