% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_ared}
\alias{ntrm_ared}
\title{Actual reduction}
\usage{
ntrm_ared(
  nu,
  f_current,
  f_candidate,
  s_current,
  s_candidate,
  m0_current,
  m0_candidate,
  mu
)
}
\arguments{
\item{nu}{penalty parameter.}

\item{f_current}{current value of the original function, i.e. \code{f(x)}.}

\item{f_candidate}{candidate value of original function, i.e. \code{f(x + p_x)}.}

\item{s_current}{current value of the slack variables.}

\item{s_candidate}{candidate value of the slack variables, i.e. \code{s + p_s}.}

\item{m0_current}{norm of \code{C(x) - s}.}

\item{m0_candidate}{norm of \code{C(x + p_x) - (s + p_s)}.}

\item{mu}{barrier penalty parameter.}
}
\value{
Value of the actual reduction \code{ared(p)} for the proposed step \code{p}.
}
\description{
Evaluate the actual reduction \code{ared(p)} for the proposed step \code{p}.
}
\details{
The actual reduction is defined as the difference of merit functions
\code{phi(x, s) - phi(x + p_x, s + p_s)} where
\verb{phi(x, s) = f(x) - mu sum(log(s)) + nu |C(x) - s|}.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
