% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_unconstrained.R
\name{ntrm_calc_p}
\alias{ntrm_calc_p}
\title{Solution to the (unconstrained) trust-region sub-problem}
\usage{
ntrm_calc_p(Q, X, L, lambda)
}
\arguments{
\item{Q}{numeric matrix of dot products between eigenvectors and gradient.}

\item{X}{numeric matrix of eigenvectors.}

\item{L}{numeric vector of eigenvalues.}

\item{lambda}{lambda parameter of the trust-region method.}
}
\value{
Current best solution to the trust-region sub-problem.
}
\description{
Compute the current best step of the trust-region sub-problem.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
