% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_constrained}
\alias{ntrm_constrained}
\title{Newton Trust-Region Interior-Point Method}
\usage{
ntrm_constrained(
  fn,
  gradient,
  hessian,
  init,
  max_iter,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{fn}{function handle to evaluate the function to minimize.}

\item{gradient}{function handle to evaluate the gradient of the function to
minimize.}

\item{hessian}{function handle to evaluate the Hessian of the function to
minimize.}

\item{init}{numeric vector of starting values.}

\item{max_iter}{integer value for the maximum number of iterations.}

\item{lower_bound}{numeric vector of lower bounds for each variable.}

\item{upper_bound}{numeric vector of upper bounds for each variable.}
}
\value{
Numeric vector \code{x} for which \code{fn(x)} is the minimum within the
boundaries specified by \code{lower_bound} and \code{upper_bound}.
}
\description{
Find the minimum of a function using the Newton method, combined with the
Trust-Region Interior-Point method.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.

Richard H. Byrd, Mary E. Hribar, and Jorge Nocedal. An interior point
algorithm for large scale nonlinear programming.
\strong{SIAM Journal on Optimization}, 9(4):877–900, 1999.
doi: 10.1137/S1052623497325107.

Todd Dennis Plantenga.
\strong{Large-scale nonlinear constrained optimization using trust regions}.
Doctoral dissertation, Northwestern University, 1994.
}
