% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_create_constraints}
\alias{ntrm_create_constraints}
\title{Process constraint requirements}
\usage{
ntrm_create_constraints(lower_bound, upper_bound)
}
\arguments{
\item{lower_bound}{numeric vector of lower bounds.}

\item{upper_bound}{numeric vector of upper bounds.}
}
\value{
A list containing the Jacobian matrix \code{A} and the indices of the
finite lower bounds \code{idx_lb} and finite upper bounds \code{idx_ub} constraints.
}
\description{
Given a set of lower bounds and upper bounds on the parameters, construct
the Jacobian of the constraint inequalities.
}
