% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_error}
\alias{ntrm_error}
\title{Maximum error of the current solution}
\usage{
ntrm_error(obj, mu)
}
\arguments{
\item{obj}{list containing all relevant variables.}

\item{mu}{barrier penalty parameter.}
}
\value{
Maximum of the three norms.
}
\description{
Compute a measure of how far the Lagrangian gradient is from zero.
}
\details{
The solution to the minimization problem is such that the gradient of our
Lagrangian function is equal to zero. After splitting the full Lagrangian
gradient into three vectors, the system of equations is the following:

G - A' z = 0
S Z - mu e = 0
C(x) - s = 0

where \code{G} is the gradient of the original function that must be minimized,
\code{C(.)} are functions evaluating the inequality constraints (in our case
they are the distances from the boundary), \code{A} is the Jacobian matrix of
\code{C(.)}, \code{z} are the Lagrangian multipliers, \code{s} are the slack variables,
\code{Z} is the diagonal matrix of Lagrange multipliers, \code{S} is the diagonal
matrix of slack variables,\code{e} is a vector of ones, and \code{mu} is a penalty
barrier parameter.

We compute the maximum norm of the previous three vectors.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
