% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_inertia_correction}
\alias{ntrm_inertia_correction}
\title{Inertia correction and regularization}
\usage{
ntrm_inertia_correction(H, A, slack_variable, lagrange_multiplier, kappa)
}
\arguments{
\item{H}{Hessian matrix.}

\item{A}{Jacobian matrix of the inequality constraints.}

\item{slack_variable}{current value of the slack variables.}

\item{lagrange_multiplier}{current value of the Lagrange multipliers.}

\item{kappa}{perturbation used in the previous interior-point iteration.}
}
\value{
Updated value of the \code{kappa} parameter.
}
\description{
Update the Hessian matrix so that the primal-dual matrix is positive
definite.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
