% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_init_obj}
\alias{ntrm_init_obj}
\title{Initialize solution object}
\usage{
ntrm_init_obj(fn, gh, C, x, s, A, mu)
}
\arguments{
\item{fn}{function handle to evaluate the function to minimize.}

\item{gh}{function handle to evaluate both the gradient and Hessian of the
function to minimize.}

\item{C}{function handle to evaluate the distances of the current solution
from the boundary.}

\item{x}{initial values of the solution.}

\item{s}{initial values of slack variables.}

\item{A}{Jacobian of the constraints}

\item{mu}{barrier penalty parameter.}
}
\value{
List with all important variables initialized.
}
\description{
Initialize solution object
}
