% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_lagrange_multiplier}
\alias{ntrm_lagrange_multiplier}
\title{Compute Lagrange multipliers}
\usage{
ntrm_lagrange_multiplier(obj, mu)
}
\arguments{
\item{obj}{list containing all the relevant variables.}

\item{mu}{barrier penalty parameter.}
}
\value{
Numeric vector of least-squares Lagrange multipliers.
}
\description{
Update the Lagrange multipliers given the current optimum and slack
variables.
}
\details{
Given the current value of the optimum and slack variables, find the
least-squares Lagrange multipliers. Define the matrix \code{B = cbind(A, -S)}
where \code{A} is the Jacobian of the constraint functions and \code{S} is the diagonal
matrix of slack variables. Define the vector \verb{u = c(G, -mu e)} where \code{G} is
the gradient of the original function and \code{e} is a vector of ones.
The least-squares multipliers are \verb{z = (B B')^(-1) B u}. If \code{z[i]} is
negative, it is substituted by the formula
\code{min(0.001, mu / slack_variable[i])}.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
