% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_functions.R
\name{ntrm_line_search}
\alias{ntrm_line_search}
\title{Find the minimum of a function by line search}
\usage{
ntrm_line_search(f, a, b)
}
\arguments{
\item{f}{one-dimensional function to minimize.}

\item{a}{lower bound.}

\item{b}{upper bound.}
}
\value{
Scalar value \verb{a <= x <= b} such that \code{f(x)} is minimum.
}
\description{
Using the Brent's method, find the value of \code{x} that minimize a function
\code{f(x)} in the interval \verb{[a, b]}.
}
\details{
This is a direct port of the algorithm from section 10.3 at page 498 of Press
et al. (2007).
}
\references{
William H. Press et al. \strong{Numerical recipes}. Cambridge University Press,
Cambridge, UK, third edition, 2007. ISBN 978-0-511-33555-6.
}
