% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_functions.R
\name{ntrm_max_abs}
\alias{ntrm_max_abs}
\title{Maximum absolute value}
\usage{
ntrm_max_abs(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Maximum absolute value in \code{x}.
}
\description{
Find the maximum absolute value in a vector, i.e. \code{max(abs(x))}, calling
\code{abs} only once.
}
