% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_merit}
\alias{ntrm_merit}
\title{Merit function}
\usage{
ntrm_merit(nu, f, s, m0, mu)
}
\arguments{
\item{nu}{penalty parameter.}

\item{f}{value of the original function \code{f(x)} we are minimizing.}

\item{s}{vector of slack variables.}

\item{m0}{norm of \code{C(x) - s}.}

\item{mu}{barrier penalty parameter.}
}
\value{
Value of the merit function evaluated at the provided parameter.
}
\description{
Evaluate the merit function
\verb{phi(x, s) = f(x) - mu sum(log(s)) + nu |C(x) - s|}.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
