% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_functions.R
\name{ntrm_norm}
\alias{ntrm_norm}
\title{Euclidean norm of a vector}
\usage{
ntrm_norm(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Euclidean norm of \code{x}.
}
\description{
Compute the Euclidean norm of a vector \code{x}, i.e. \verb{|x| = sqrt(sum(x^2))}.
}
