% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_objective_function}
\alias{ntrm_objective_function}
\title{Objective function value}
\usage{
ntrm_objective_function(p, Q, u)
}
\arguments{
\item{p}{numeric vector at which the function is to be evaluated.}

\item{Q}{numeric matrix representing the quadratic part.}

\item{u}{numeric vector representing the linear part.}
}
\value{
Value of the function.
}
\description{
Compute the value of the sub-problem objective function.
}
\details{
The function evaluated here is

\verb{h(p) = 0.5 p' Q p + u' p}
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
