% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_pred}
\alias{ntrm_pred}
\title{Predicted reduction}
\usage{
ntrm_pred(nu, m0, mp, h)
}
\arguments{
\item{nu}{penalty parameter.}

\item{m0}{norm of \code{C(x) - s}.}

\item{mp}{norm of \verb{A p_x - p_s + C(x) - s}.}

\item{h}{value of the function \code{h(p_x, p_s)} (see details).}
}
\value{
Value of the predicted reduction \code{pred(p)} for the proposed step \code{p}.
}
\description{
Evaluate the predicted reduction \code{pred(p)} for the proposed step \code{p}.
}
\details{
The predicted reduction is defined as the difference \code{q(0, nu) - q(p, nu)}
where \verb{q(p, nu) = h(p_x, p_s) + nu m(p)}, \code{h(p_x, p_s)} is equal to
\verb{0.5 (p_x' H p_x + p_s' S^(-1) Z p_s) + (G' p_x - mu e' S^(-1) p_s)}, and
\verb{m(p) = |A p_x - p_s + C(x) - s|}. \code{G} is the gradient of the original
function that must be minimized, \code{C(.)} are functions evaluating the
inequality constraints (in our case they are the distances from the
boundary), \code{A} is the Jacobian matrix of \code{C(.)}, \code{z} are the Lagrangian
multipliers, \code{s} are the slack variables, \code{Z} is the diagonal matrix of
Lagrange multipliers, \code{S} is the diagonal matrix of slack variables,\code{e} is a
vector of ones, and \code{nu} and \code{mu} are penalty parameters.

Note that \code{pred(p)} is simply equal to \code{nu (m(0) - m(p)) - h(p_x, p_s)}.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
