% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_reductions}
\alias{ntrm_reductions}
\title{Evaluate the proposed step}
\usage{
ntrm_reductions(obj, candidate, mu, nu)
}
\arguments{
\item{obj}{list containing all relevant variables of the current best
solution.}

\item{candidate}{list containing all relevant variables of the proposed new
solution.}

\item{mu}{barrier penalty parameter.}

\item{nu}{constraints penalty parameter.}
}
\value{
A list containing the updated parameter \code{nu}, the actual reduction,
the predicted reduction, and their ratio.
}
\description{
Compute the actual reduction, predicted reduction, and their ratio.
}
\details{
The actual reduction is defined as the difference of merit functions
\code{phi(x, s) - phi(x + p_x, s + p_s)} where
\verb{phi(x, s) = f(x) - mu sum(log(s)) + nu |C(x) - s|}.
The predicted reduction is instead simply equal to
\code{nu (m(0) - m(p)) - h(p_x, p_s)}.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
