% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm.R
\name{ntrm_safeguard}
\alias{ntrm_safeguard}
\title{Safeguard parameter computation}
\usage{
ntrm_safeguard(lambda, G, H, delta)
}
\arguments{
\item{lambda}{lambda parameter of the trust-region method.}

\item{G}{gradient vector.}

\item{H}{hessian matrix.}

\item{delta}{radius of the trust region.}
}
\value{
Updated scalar value \code{lambda}.
}
\description{
Update the current value of \code{lambda} to hopefully obtain a non-singular
matrix.
}
\references{
Jorge J Moré and D C Sorensen. Computing a Trust Region Step.
\strong{SIAM Journal on Scientific and Statistical Computing}, 4(3):553-572, 1983.
doi: 10.1137/0904038.
}
