% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_functions.R
\name{ntrm_solve_quadratic_equation}
\alias{ntrm_solve_quadratic_equation}
\title{Solutions of a quadratic equation}
\usage{
ntrm_solve_quadratic_equation(a, b, c)
}
\arguments{
\item{a}{quadratic coefficient.}

\item{b}{linear coefficient.}

\item{c}{constant.}
}
\value{
Numeric vector of length 2 with the solutions of the quadratic
equation.
}
\description{
Solve the equation \verb{a x^2 + b x + c = 0} for x, in a numerically stable
manner. However, the discriminant is still evaluated using the actual
formula.
}
