% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_unconstrained.R
\name{ntrm_solve_tr_subproblem}
\alias{ntrm_solve_tr_subproblem}
\title{Iterative solution to the (unconstrained) trust-region sub-problem}
\usage{
ntrm_solve_tr_subproblem(G, H, delta)
}
\arguments{
\item{G}{gradient vector.}

\item{H}{hessian matrix.}

\item{delta}{radius of the trust region.}
}
\value{
List with the best step solution \code{p} to the trust-region sub-problem
and the value \code{m} of the objective function \code{m(p)}.
}
\description{
Compute the best step of the trust-region sub-problem.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
