% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_step}
\alias{ntrm_step}
\title{Step computation}
\usage{
ntrm_step(obj, delta, tau)
}
\arguments{
\item{obj}{list containing all relevant variables.}

\item{delta}{radius of the trust region.}

\item{tau}{fraction to the boundary parameter.}
}
\value{
List with the step of the interior point trust-region method,
satisfying the boundary constraints.
}
\description{
Compute the step of the interior point trust-region method
}
\details{
This is the core function of the interior point trust-region method. We seek
the vector \verb{p = (p_x, p_s)} minimizing
\verb{h(p_x, p_s) = G' p_x + 0.5 p_x' H p_x - mu e' p_s + 0.5 p_s' Z S p_s}
subject to the constraints \verb{A p_x - S p_s + (C(x) - s) = k}, \verb{|p| <= delta},
and \code{min(p_s) >= -tau}.

\code{G} is the gradient of the original function, \code{H} is the Hessian matrix of
the Lagrangian function, \code{mu} is the barrier penalty parameter, \code{Z} is the
diagonal matrix of Lagrange multipliers, \code{S} is the diagonal matrix of slack
variables \code{s}, \code{C(.)} are functions evaluating the inequality constraints
(in our case they are the distances from the boundary), \code{A} is the
Jacobian matrix of \code{C(.)}, and \code{e} is a vector of ones.

Our objective function is made of two separate quadratic programs. The first
part is associated with the original variables:

QP1(p_x) = 0.5 p_x' H p_x + G' p_x

The second part is associated with the slack variables:

QP2(p_s) = 0.5 p_s' V p_s - mu' p_s

where \verb{V = Z S}.

Set \code{w(x) = C(x) - s}, \code{p = c(p_x, p_s)},
\code{Q = rbind(cbind(H, 0), cbind(0, V))}, \code{B = cbind(A, -S)}, \verb{u = c(G, -mu e)},
\code{y(x) = k - w(x)}. The problem becomes:

min_{p} 0.5 p' Q p + p' u
subject to B p = y(x)

which we can solve with the projected conjugate gradient method.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
