% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_step_normal}
\alias{ntrm_step_normal}
\title{Normal step computation}
\usage{
ntrm_step_normal(Y, B, slack_residual, delta, tau, idx_s)
}
\arguments{
\item{Y}{numeric matrix used to solve the linear system.}

\item{B}{numeric matrix defined as \code{B = cbind(A, -S)}.}

\item{slack_residual}{numeric vector of distances from the boundary.}

\item{delta}{radius of the trust region.}

\item{tau}{fraction to the boundary parameter.}

\item{idx_s}{integer vector indicating the position of the slack variables in
the \code{initial_value} starting vector.}
}
\value{
Numeric vector approximately minimizing \verb{|B v + C(x) - s|^2} subject
to the constraints.
}
\description{
Compute the normal step of the interior point trust-region method.
}
\details{
We want to find the minimum of \verb{(B v + w)'(B v + w)} where \code{B = cbind(A, -S)}
and \code{w = C(x) - s}. \code{C(.)} are functions evaluating the inequality
constraints (in our case they are the distances from the boundary), \code{A} is
the Jacobian matrix of \code{C(.)}, \code{s} are the slack variables, and \code{S} is the
diagonal matrix of slack variables.

The solution must also satisfy the constraints \verb{|v| <= 0.8 delta} and
\code{min(v_s) >= -tau / 2}, where \code{v_s} are the values of \code{v} corresponding to
the slack variables.

The solution of the problem is computed with the dogleg method.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
