% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_step_shrink}
\alias{ntrm_step_shrink}
\title{Shrink a candidate step}
\usage{
ntrm_step_shrink(p, a, d, tau, idx_s)
}
\arguments{
\item{p}{previous step in the iteration.}

\item{a}{magnitude of the new proposed direction move.}

\item{d}{direction towards the new step solution.}

\item{tau}{fraction to the boundary parameter.}

\item{idx_s}{integer vector indicating the position of the slack variables in
the vector \code{p}.}
}
\value{
New step obtained by starting at \code{p} and moving along direction \code{d}
such that \code{min(p_new[idx_s]) >= -tau}.
}
\description{
Shrink a candidate step so that \code{min(p_s) >= -tau}.
}
