% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_step_tilde}
\alias{ntrm_step_tilde}
\title{Step computation}
\usage{
ntrm_step_tilde(Q, u, B, initial_value, delta, tau, idx_s)
}
\arguments{
\item{Q}{numeric matrix representing the quadratic part.}

\item{u}{numeric vector representing the linear part.}

\item{B}{numeric matrix representing the linear constraints.}

\item{initial_value}{initial vector satisfying the equality constraints.}

\item{delta}{radius of the trust region.}

\item{tau}{fraction to the boundary parameter.}

\item{idx_s}{integer vector indicating the position of the slack variables in
the \code{initial_value} starting vector.}
}
\value{
List containing the step of the interior point trust-region method
satisfying the boundary constraints, and the minimum value of the objective
function.
}
\description{
Apply the projected gradient method to find the candidate step of the
interior point trust-region method.
}
\details{
The problem we want to solve is

min_{p} 0.5 p' Q p + p' u
subject to B p = y(x)

If the initial value satisfies the linear constraints, the projected gradient
method is guaranteed to automatically iterate among solutions that also
satisfy the constraints.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.
}
