% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_update_nu}
\alias{ntrm_update_nu}
\title{Update penalty parameter}
\usage{
ntrm_update_nu(nu, h, m0, mp)
}
\arguments{
\item{nu}{current value of the penalty parameter.}

\item{h}{value of the interior point trust-region objective function.}

\item{m0}{norm of \code{C(x) - s}.}

\item{mp}{norm of \verb{A p_x - p_s + C(x) - s}.}
}
\value{
Updated penalty parameter such that \verb{pred(d) >= 0.3 nu (m0 - mp)}.
}
\description{
Update the penalty parameter of the merit function.
}
\references{
Jorge Nocedal and Stephen J Wright. \strong{Numerical optimization}. Springer,
New York, NY, USA, second edition, 2006. ISBN 978-0-387-30303-1.

Richard H. Byrd, Mary E. Hribar, and Jorge Nocedal. An interior point
algorithm for large scale nonlinear programming.
\strong{SIAM Journal on Optimization}, 9(4):877–900, 1999.
doi: 10.1137/S1052623497325107.
}
