% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrm_constrained.R
\name{ntrm_update_solution}
\alias{ntrm_update_solution}
\title{Update the solution}
\usage{
ntrm_update_solution(obj, candidate, gh, mu)
}
\arguments{
\item{obj}{list containing all relevant variables of the current best
solution.}

\item{candidate}{list containing all relevant variables of the proposed new
solution.}

\item{gh}{function handle to evaluate both the gradient and Hessian of the
function to minimize.}

\item{mu}{barrier penalty parameter.}
}
\description{
Update the old solution with a better one.
}
