% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.drda}
\alias{plot.drda}
\title{Model fit plotting}
\usage{
\method{plot}{drda}(x, ...)
}
\arguments{
\item{x}{\code{drda} object as returned by the \code{link[drda]{drda}} function.}

\item{...}{other \code{drda} objects or parameters to be passed to the plotting
functions. See 'Details'.}
}
\value{
If argument \code{show} is \code{TRUE}, no return value. If argument \code{show} is
\code{FALSE}, a list with all plotting data.
}
\description{
Plot maximum likelihood curves fitted with \code{drda}.
}
\details{
This function provides a scatter plot of the observed data, overlaid with the
maximum likelihood curve fit.
If multiple fit objects from the same family of models are given, they will
all be placed in the same plot.

Accepted plotting arguments are:

\describe{
\item{base}{character string with the base used for printing the values on
the \code{x} axis. Accepted values are \code{10} for base 10, \code{2} for base 2, \code{e}
for base e, or \code{n} (default) for no log-scale printing.}
\item{col}{curve color(s). By default, up to 9 color-blind friendly
colors are provided.}
\item{xlab, ylab}{axis labels.}
\item{xlim, ylim}{the range of x and y values with sensible defaults.}
\item{level}{level of confidence intervals (default is 0.95). Set to zero
or a negative value to disable confidence intervals.}
\item{midpoint}{if \code{TRUE} (default) shows guidelines associated with the
curve mid-point.}
\item{plot_data}{if \code{TRUE} (default) shows in the plot the data points used
for fitting.}
\item{legend_show}{if \code{TRUE} (default) shows the legend.}
\item{legend_location}{character string with custom legend position. See
\code{link[graphics]{legend}} for possible keywords.}
\item{legend}{custom labels for the legend model names.}
\item{show}{If \code{TRUE} (default) a figure is plotted, otherwise the function
returns a list with values to create the figure manually.}
}
}
