% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{rss.logistic4}
\alias{rss.logistic4}
\alias{rss_fixed.logistic4}
\title{Residual sum of squares}
\usage{
\method{rss}{logistic4}(object)

\method{rss_fixed}{logistic4}(object, known_param)
}
\arguments{
\item{object}{object of class \code{logistic4}.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any.}
}
\value{
Function handle \code{f(theta)} to evaluate the RSS associated to a
particular parameter choice \code{theta}.
}
\description{
Evaluate the residual sum of squares (RSS) against the mean of a
4-parameter logistic model.
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + omega / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, omega, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{omega} is the width of the curve, \code{eta} is the steepness
of the curve or growth rate (also known as the Hill coefficient), and \code{phi}
is the value of \code{x} at which the curve is equal to its mid-point.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega') / (1 + exp(-eta * (x - phi)))}
}
