% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{rss_gradient_hessian.gompertz}
\alias{rss_gradient_hessian.gompertz}
\alias{rss_gradient_hessian_fixed.gompertz}
\title{Residual sum of squares}
\usage{
\method{rss_gradient_hessian}{gompertz}(object)

\method{rss_gradient_hessian_fixed}{gompertz}(object, known_param)
}
\arguments{
\item{object}{object of class \code{gompertz}.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any.}
}
\value{
Function handle \code{f(theta)} to evaluate the gradient and Hessian of
the RSS associated to a particular parameter choice \code{theta}.
}
\description{
Evaluate the gradient and Hessian of the residual sum of squares (RSS)
against the mean of a Gompertz model.
}
\details{
The Gompertz function \verb{f(x; theta)} is defined here as

\verb{alpha + (beta - alpha) exp(-exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate, and \code{phi} related to the function
value at \code{x = 0}.
}
