% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{rss_gradient_hessian.logistic4}
\alias{rss_gradient_hessian.logistic4}
\alias{rss_gradient_hessian_fixed.logistic4}
\title{Residual sum of squares}
\usage{
\method{rss_gradient_hessian}{logistic4}(object)

\method{rss_gradient_hessian_fixed}{logistic4}(object, known_param)
}
\arguments{
\item{object}{object of class \code{logistic4}.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any.}
}
\value{
Function handle \code{f(theta)} to evaluate the gradient and Hessian of
the RSS associated to a particular parameter choice \code{theta}.
}
\description{
Evaluate the gradient and Hessian of the residual sum of squares (RSS)
against the mean of a 4-parameter logistic model.
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined here as

\code{alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate (also known as the Hill coefficient),
and \code{phi} is the value of \code{x} at which the curve is equal to its mid-point.
}
