% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{variance_normal}
\alias{variance_normal}
\title{Variance estimator}
\usage{
variance_normal(rss, df)
}
\arguments{
\item{rss}{value of the residual sum of squares.}

\item{df}{residual degrees of freedom.}
}
\value{
Corrected maximum likelihood estimate of the variance.
}
\description{
Compute the corrected variance estimate associated with a Normal
distribution.
}
