% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voropm2}
\alias{voropm2}
\title{Vorinostat in OPM-2 cell-line dataset}
\format{
A data frame with 45 rows and 4 variables:
\describe{
\item{response}{viability measures normalized using positive and
negative controls}
\item{dose}{drug concentrations (nM) used for testing}
\item{log_dose}{natural logarithm of variable \code{dose}}
\item{weight}{random weights included only for package demonstration}
}
}
\usage{
voropm2
}
\description{
A dataset containing dose-response data of drug Vorinostat tested ex-vivo on
the OPM-2 cell-line.
}
\keyword{datasets}
