% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda-package.R
\docType{package}
\name{drda_package}
\alias{drda_package}
\title{Dose-response data analysis}
\description{
Description.
}
\section{Available functions}{


\itemize{
\item{\code{\link{drda}}: logistic fit.}
}
}

\references{
Author et al. (2020).
}
