% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic4.R
\name{gradient_hessian.logistic4}
\alias{gradient_hessian.logistic4}
\title{4-parameter logistic function}
\usage{
\method{gradient_hessian}{logistic4}(object, theta)
}
\arguments{
\item{object}{object of class \code{logistic4}.}

\item{theta}{numeric vector with the four parameters in the form
\code{c(alpha, beta, eta, phi)}.}
}
\value{
List of two elements: \code{G} the gradient and \code{H} the Hessian.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
4-parameter logistic function.
}
\details{
The 4-parameter logistic function \verb{f(x; theta)} is defined here as

\code{alpha + (beta - alpha) / (1 + exp(-eta * (x - phi)))}

where \code{theta = c(alpha, beta, eta, phi)}, \code{alpha} is the lower horizontal
asymptote, \code{beta > alpha} is the upper horizontal asymptote, \code{eta} is the
steepness of the curve or growth rate (also known as the Hill coefficient),
and \code{phi} is the value of \code{x} at which the curve is equal to its mid-point.
}
