% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{gradient_hessian.logistic5}
\alias{gradient_hessian.logistic5}
\title{5-parameter logistic function}
\usage{
\method{gradient_hessian}{logistic5}(object, theta)
}
\arguments{
\item{object}{object of class \code{logistic5}.}

\item{theta}{numeric vector with the five parameters in the form
\code{c(alpha, beta, eta, phi, nu)}.}
}
\value{
List of two elements: \code{G} the gradient and \code{H} the Hessian.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
5-parameter logistic function.
}
\details{
The 5-parameter logistic function \verb{f(x; theta)} is defined here as

\code{alpha + (beta - alpha) / (1 + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu)}, \code{beta > alpha}, and \code{nu > 0}.
}
