% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic6.R
\name{gradient_hessian.logistic6}
\alias{gradient_hessian.logistic6}
\title{6-parameter logistic function}
\usage{
\method{gradient_hessian}{logistic6}(object, theta)
}
\arguments{
\item{object}{object of class \code{logistic6}.}

\item{theta}{numeric vector with the six parameters in the form
\code{c(alpha, beta, eta, phi, nu, xi)}.}
}
\value{
List of two elements: \code{G} the gradient and \code{H} the Hessian.
}
\description{
Evaluate at a particular set of parameters the gradient and Hessian of the
6-parameter logistic function.
}
\details{
The 6-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (xi + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu, xi)}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega) / (exp(xi') + exp(nu' - eta * (x - phi)))^(-exp(-nu'))}

Note that it is simply \code{beta = alpha + exp(omega)}, \verb{nu = exp(nu')},
\verb{xi = exp(xi')}.

NOTE: This function expects the input vector
\code{theta = c(alpha, beta, eta, phi, nu, xi)} but computes the gradient and
Hessian of the alternative parametrization
\verb{theta' = (alpha, omega, eta, phi, nu', xi')}.
This duality is unfortunately a trade-off between what users expect (first
version is easier to interpret) and the optimization stability of the second
version.
}
