% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic5.R
\name{rss.logistic5}
\alias{rss.logistic5}
\alias{rss_fixed.logistic5}
\title{Residual sum of squares}
\usage{
\method{rss}{logistic5}(object)

\method{rss_fixed}{logistic5}(object, known_param)
}
\arguments{
\item{object}{object of class \code{logistic5}.}

\item{known_param}{numeric vector with the known fixed values of the model
parameters, if any.}
}
\value{
Function handle \code{f(theta)} to evaluate the RSS associated to a
particular parameter choice \code{theta}.
}
\description{
Evaluate the residual sum of squares (RSS) against the mean of a
5-parameter logistic model.
}
\details{
The 5-parameter logistic function \verb{f(x; theta)} is defined in this package as

\code{alpha + (beta - alpha) / (1 + nu * exp(-eta * (x - phi)))^(1 / nu)}

where \code{theta = c(alpha, beta, eta, phi, nu)}.

In our optimization algorithm, however, we consider instead the equivalent
function \verb{f(x; theta')}

\verb{alpha + exp(omega) / (1 + exp(nu' - eta * (x - phi)))^(-exp(-nu'))}

Note that it is simply \code{beta = alpha + exp(omega)} and \verb{nu = exp(nu')}.
}
