% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{suff_stats}
\alias{suff_stats}
\title{Sufficient statistics}
\usage{
suff_stats(x, y, w)
}
\arguments{
\item{x}{numeric vector representing the fixed predictor variable.}

\item{y}{numeric vector of observed values.}

\item{w}{numeric vector of weights.}
}
\value{
Numeric matrix where the first column are the sorted unique values of
input vector \code{x}, second column are the total weights (sample size if no
weights), third column are the (weighted) sample means, and fourth column
are the uncorrected (weighted) sample variances.
}
\description{
Compute sufficient statistics for each unique value of the predictor
variable.
}
